+++ include/CommonTypes.hpp
@@ -29,6 +29,7 @@ enum class NetMuxPacketType {
     Ping,
     InputEvent,
     ResolutionUpdate,
+    WebRTCSignaling,
     AuthChallenge,
     AuthResponse,
     SyncCheck,