+++ include/DesktopCapture.hpp
@@ -5,6 +5,9 @@
 #ifdef _WIN32
 #include <d3d11.h>
 #include <dxgi1_2.h>
+#else
+#include <X11/Xlib.h>
+#include <X11/Xutil.h>
 #endif
 
 class DesktopCapture {
@@ -18,6 +21,8 @@ public:
 
 #ifdef _WIN32
     ID3D11Texture2D* GetCurrentFrameTexture() { return m_currentFrame; }
+#else
+    XImage* GetCurrentFrameImage() { return m_currentImage; }
 #endif
 
 private:
@@ -26,5 +31,11 @@ private:
     ID3D11DeviceContext* m_context;
     IDXGIOutputDuplication* m_deskDupl;
     ID3D11Texture2D* m_currentFrame;
+#else
+    Display* m_display;
+    Window m_rootWindow;
+    XImage* m_currentImage;
+    int m_screenWidth;
+    int m_screenHeight;
 #endif
 };