+++ include/NetMuxFramework.hpp
@@ -97,6 +97,7 @@ private:
     double m_lastSyncTime;
     double m_lastPerfLog;
     bool m_overlayDirty;
+    bool m_webrtcInitialized;
 
     struct InteractionEvent {
         unsigned long long peerId;