+++ include/SpatialViewport.hpp
@@ -29,8 +29,16 @@ private:
     DirectX::XMMATRIX m_projMatrix;
     float m_currentCamX;
     float m_currentCamZ;
+    float m_transitionProgress;
 
     ID3D11ShaderResourceView* m_localSRV;
     ID3D11ShaderResourceView* m_remoteSRV;
+
+    ID3D11Buffer* m_vertexBuffer;
+    ID3D11Buffer* m_constantBuffer;
+    ID3D11VertexShader* m_vertexShader;
+    ID3D11PixelShader* m_pixelShader;
+    ID3D11InputLayout* m_inputLayout;
+    ID3D11SamplerState* m_samplerState;
 #endif
 };